+++ tests/unit_tests/storage/test_local_ensemble.py
@@ -20,8 +20,8 @@ def test_that_egrid_files_are_saved_and_loaded_correctly(tmp_path):
         ensemble_dir = tmp_path / "ensembles" / str(ensemble.id)
         assert ensemble_dir.exists()
 
-        data = np.full_like(mask_values, np.nan, dtype=np.double)
-        np.place(data, mask_values, np.array([1.2, 1.1, 4.3, 3.1]))
+        data = np.full_like(mask_values, np.nan, dtype=np.float32)
+        np.place(data, mask_values, np.array([1.2, 1.1, 4.3, 3.1], dtype=np.float32))
         da = xr.DataArray(
             data.reshape((4, 5, 1)), name="values", dims=["x", "y", "z"]  # type: ignore
         )
@@ -43,8 +43,8 @@ def test_that_grid_files_are_saved_and_loaded_correctly(tmp_path):
         grid = EclGridGenerator.create_rectangular((4, 5, 1), (1, 1, 1), actnum=mask)
         grid.save_GRID(f"{experiment.mount_point}/grid.GRID")
 
-        data = np.full_like(mask, np.nan, dtype=np.double)
-        np.place(data, mask, np.array([1.2, 1.1, 4.3, 3.1]))
+        data = np.full_like(mask, np.nan, dtype=np.float32)
+        np.place(data, mask, np.array([1.2, 1.1, 4.3, 3.1], dtype=np.float32))
         da = xr.DataArray(
             data.reshape((4, 5, 1)), name="values", dims=["x", "y", "z"]  # type: ignore
         )
